ThingSet interfaces merged with Ext interfaces
