--- src/ext/thingset_interface.h
@@ -1,44 +1,0 @@
-/* LibreSolar charge controller firmware
- * Copyright (c) 2016-2019 Martin Jäger (www.libre.solar)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#ifndef THINGSET_INTERFACE_H
-#define THINGSET_INTERFACE_H
-
-#include <vector>
-
-class ThingSetInterface
-{
-    public:
-        virtual void process_asap()  {};
-        virtual void process_1s() {};
-        virtual void enable() {};
-    private:
-
-};
-
-class ThingSetInterfaceManager: public ThingSetInterface
-{
-    public:
-        virtual void process_asap();
-        virtual void process_1s();
-        virtual void enable();
-
-    private:
-        static std::vector<ThingSetInterface*> interfaces;
-};
-
-extern ThingSetInterfaceManager ts_interfaces;
-#endif
